fix: support arrays
